fix: reverted Site to Provider
The name Site was being used twice in spot.ts

Also removed the "provider" attribute from Collection, since it was not
being used.

Removed a few of the example Providers from spot.ts to make the code a
bit simpler.

Modified some code to make it lopp instead of functional. This is to
help with later refactoring.